Revision: 34295

+++ SequoiaDB/engine/cls/clsVSSecondary.cpp
@@ -60,7 +60,7 @@ namespace engine
    {
       PD_TRACE_ENTRY ( SDB__CLSVSSD_HDINPUT ) ;
       SDB_ASSERT( NULL != header, "header should not be NULL" ) ;
-      if ( MSG_CLS_BALLOT_RES == (UINT32)header->opCode )
+      if ( MSG_CLS_BALLOT_RES == header->opCode )
       {
          next = id() ;
          goto done ;
@@ -90,7 +90,6 @@ namespace engine
       }
       else
       {
-         /// error msg
          next = id() ;
       }
    done:
